+++ app/global.css
@@ -6,21 +6,21 @@
 
 @font-face {
   font-family: 'BahijTheSansArabic';
-  src: url('../../public/fonts/Bahij_TheSansArabic-Bold.ttf') format('truetype');
+  src: url('../public/fonts/Bahij_TheSansArabic-Bold.ttf') format('truetype');
   font-weight: 900;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'BahijTheSansArabic';
-  src: url('../../public/fonts/Bahij_TheSansArabic-Plain.ttf') format('truetype');
+  src: url('../public/fonts/Bahij_TheSansArabic-Plain.ttf') format('truetype');
   font-weight: 400;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'BahijTheSansArabic';
-  src: url('../../public/fonts/Bahij_TheSansArabic-SemiBold.ttf') format('truetype');
+  src: url('../public/fonts/Bahij_TheSansArabic-SemiBold.ttf') format('truetype');
   font-weight: 600;
   font-style: normal;
 }
